Use templates and if constexpr for a more optimised code path
By making save_anim into a template argument, rather than a function
argument, we can generate two different functions where this argument
is known at compile time. This allows us to use `if constexpr`, which
will allow the compiler to evaluate the expression and optimise out
the unnecessary branches. This should make for a cleaner code path
when animations aren't being saved. We can verify that this is the
case by looking at the decompilation of a debug build in Ghidra; the
`delaunay_triangulate<false>` function has no calls to
`save_triangulation`, whereas `delaunay_triangulate<true>` has
unconditional calls to `save_triangulation`. Maybe this is an
unnecessary optimisation that makes the binary less space-efficient,
but the code paths in this instance are much clearer with reduced branching.

Also use this for deciding whether or not to pre-calculate the bounds
of the data, which will save time when it comes to actually saving the frames.